requirements.txt
# FastAPI backend
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# Groq LLM
groq==0.9.0

# AI / ML
torch==2.3.0
transformers==4.41.0
faiss-cpu==1.8.0
numpy==1.26.4
sentence-transformers==3.0.1

# Code parsing
tree-sitter==0.22.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
